refactor(logger): curlify log output
Updates the logger to resemble more curl-esque requests and responses
(direction denoted with `>` for request and `<` for responses) when
logging the HTTP transactions.

As part of this refactor, I've also introduced a redaction for any
sensitive HTTP headers where we don't need the value and providing it in
debug logs has the potential to leak credentials.